+++ CHANGELOG.md
@@ -13,6 +13,8 @@
     - more coherent JS environment for developpers
     - lighter assets
     - drop Handlebars dependency
+- Themes are now entrypoint-based [#829](https://github.com/opendatateam/udata/pull/829).
+  There is also a new [theming documention](https://udata.readthedocs.io/en/stable/creating-theme/).
 
 ## 1.0.5 (2017-03-27)